vmm: move `KvmContext` to vstate file
KvmContext is actually coupled to the
functonality defined in vstate since it
all handles kvm interaction. Move it there and
make the appropriate changes.
